rename method and move a test thats relevant for migrate pkg
